+++ test-vectors/history/.gitignore
@@ -1,2 +1,3 @@
 /types.hpp
 /types.scale.hpp
+/types.diff.hpp
